refactor(queue): extract lock and jsonl store
